+++ app_configs/__init__.py
@@ -1,8 +1,8 @@
-from dir_maker import Config as Directories
-from user_settings import Config as UserConfig
-from keybinds import Config as KeybindConfig
-from sidebar_loader import Config as SidebarLoader
-from terminal_input_handler import Config as TerminalConfig
+from .dir_maker import Config as Directories
+from .user_settings import Config as UserConfig
+from .keybinds import Config as KeybindConfig
+from .sidebar_loader import Config as SidebarLoader
+from .terminal_input_handler import Config as TerminalConfig
 
 
 class AppConfigHandler(object):