+++ app_configs/keybinds.py
@@ -1,40 +1,45 @@
-from config_base import ConfigBase
-from app_modules.key_binder.key_binder import KeyBinder
+from app_modules.key_binder import key_binder as kbinder
+from .config_base import ConfigBase
+from kivy.app import App
 
 
 class Config(ConfigBase):
 
     @staticmethod
     def load_before(root):
-        root.keybinder = KeyBinder()
-        # root.keybinder.log_keys = True
+        pass
 
     @staticmethod
     def load_after(root):
-        root.keybinder.add(
+        app = App.get_running_app()
+        kbinder.add(
+            'glb_escape', 27, 'down',
+            app.stop, category='globhandler')
+
+        kbinder.add(
             'vol_increase', '273', 'down',
             root.ids.playback_bar.volume_increase, modifier=['ctrl'])
-        root.keybinder.add(
+        kbinder.add(
             'vol_decrease', '274', 'down',
             root.ids.playback_bar.volume_decrease, modifier=['ctrl'])
-        root.keybinder.add(
+        kbinder.add(
             'seek_4_sec_back', '276', 'down',
             lambda: root.mPlayer.seek_relative(-4), modifier=['shift'])
-        root.keybinder.add(
+        kbinder.add(
             'seek_4_sec_forward', '275', 'down',
             lambda: root.mPlayer.seek_relative(4), modifier=['shift'])
-        root.keybinder.add(
+        kbinder.add(
             'seek_60_sec_back', '276', 'down',
             lambda: root.mPlayer.seek_relative(-60), modifier=['ctrl'])
-        root.keybinder.add(
+        kbinder.add(
             'seek_60_sec_forward', '275', 'down',
             lambda: root.mPlayer.seek_relative(60), modifier=['ctrl'])
-        root.keybinder.add(
+        kbinder.add(
             'play_pause_toggle', '32', 'down', root.media_control.play_pause)
 
-        root.keybinder.add('toggle_terminal', '96', 'down',
+        kbinder.add('toggle_terminal', '96', 'down',
                            root.ids.terminal_widget.toggle_pos_multiplier)
-        root.keybinder.add('terminal_scroll_up', '280', 'down',
+        kbinder.add('terminal_scroll_up', '280', 'down',
                           root.ids.terminal_widget.scroll_up)
-        root.keybinder.add('terminal_scroll_down', '281', 'down',
+        kbinder.add('terminal_scroll_down', '281', 'down',
                         root.ids.terminal_widget.scroll_down)