+++ app_modules/media_player/providers/__init__.py
@@ -1,6 +1,6 @@
 from kivy.utils import platform
-from audio_player_kivy import AudioPlayer
-from video_player_kivy import AppVideoPlayer
+from .audio_player_kivy import AudioPlayer
+from .video_player_kivy import AppVideoPlayer
 
 
 if platform == 'android':