+++ app_modules/widgets_standalone/info_widget.py
@@ -17,13 +17,13 @@ class HoverBehavior(Widget):
         super(HoverBehavior, self).__init__(**kwargs)
         Window.bind(mouse_pos=self.on_mouse_move)
 
-    def on_mouse_move(self, win, (posx, posy)):
+    def on_mouse_move(self, win, pos):
         if self.hovering == False:
-            if self.collide_point_window(posx, posy):
+            if self.collide_point_window(*pos):
                 self.hovering = True
                 self.on_enter()
         else:
-            if self.collide_point_window(posx, posy) == False:
+            if self.collide_point_window(*pos) == False:
                 self.hovering = False
                 self.on_leave()
 
@@ -99,7 +99,7 @@ class InfoLabel(HoverBehavior, Label):
         if self.parent:
             self.parent.remove_widget(self)
         else:
-            print 'NO PARENT ', self
+            print ('NO PARENT ', self)
 
     def on_enter(self):
         self.bold = True